processor/requirements.txt
flask>=3.0
gunicorn>=22.0
google-cloud-pubsub>=2.23
google-cloud-firestore>=2.19
pytest>=8.0
